Rename suggest_wrapper to hparams.
