+++ docs/source/contribution_guide/tests.md
@@ -4,7 +4,7 @@
 
 - aiaccel uses pytest for testing.
 - Create a directory for unit test under tests directory.
-  - The directory structure under `aiaccel/tests` corresponds to that under `aiaccel/aiaccel`, except for a few modeules such as config.py. For example, the test for `aiaccel/aiaccel/hpo/optuna/suggest_wrapper.py` is `aiaccel/tests/hpo/optuna/test_suggest_wrapper.py`.
+  - The directory structure under `aiaccel/tests` corresponds to that under `aiaccel/aiaccel`, except for a few modeules such as config.py. For example, the test for `aiaccel/aiaccel/hpo/optuna/hparams.py` is `aiaccel/tests/hpo/optuna/test_hparams.py`.
 - If you have added a new feature or bug fix, please create codes for testing.
 
 
@@ -20,7 +20,7 @@ pytest
 - Specify a file name as an argument to run only specific code.
 
 ~~~bash
-pytest aiaccel/tests/hpo/optuna/test_suggest_wrapper.py
+pytest aiaccel/tests/hpo/optuna/test_hparams.py
 ~~~
 
 - In addition, execute the following command to check coding styles.
