+++ tests/hpo/apps/in_memory.yaml
@@ -12,7 +12,7 @@ params:
   _target_: aiaccel.hpo.optuna.hparams_manager.HparamsManager
   x1: [0, 1]
   x2:
-    _target_: aiaccel.hpo.optuna.suggest_wrapper.Float
+    _target_: aiaccel.hpo.optuna.hparams.Float
     name: x2
     low: 0.0
     high: 1.0
